Initial commit with first pass at API
Some parts of this are missing (eg. propagation, baggage, etc) and some
parts of this are implemented only as very naive first-approaches to
make the wiring match in places (eg. context stack).

Details about the interface, package layout, method and sub names,
general style, etc, are all considered to be up for discussion.