Add test coverage for implicit context management
Some of the code in the root OpenTelemetry module is moved in this
patch to OpenTelemetry::Common to avoid import cycles and prevent
compilation failures. This common package continues to be for internal
use only.

This patch also includes some initial tests of the logging / error
handling mechanism, which is currently implemented using Log::Any.
Whether this implementation survives in time or not remains to be seen,
but at this point we are mostly concerned with implementing the API, and
ensuring it behaves as expected. Anything not on the OpenTelemetry spec
should be considered an implementation detail, unless otherwise noted.