Expose writable attribute role as separate one
Classes with attributes should, in general, be read-only to simplify
their use across async code. Shared attribute behaviours are implemented
in a shared role, but since we currently do not have the ability to
override role methods, this patch separates the roles into a writable
one, and a non-writable one.

We still need to expose a _set_attribute method in the writable case
because we need to be able in some cases to add custom behaviour
(eg. disabling writing when a span is ended). This might change in the
future.