Move processor and exporter roles to API distribution
Although these are strictly speaking SDK concepts, if these existed
in the SDK distribution it would mean that processors and exporters
provided by third parties would have to depend on the API _and_ the
SDK, which would be unfortunate. Specially if there is a possibility
in the future for multiple SDKs to exist.