Remove writable attribute role
Downstream uses of this role that require writing can use the MOP
to fetch the role's attribute and call `set` on it. The ugliness
about it should be enough of a marker that this is not entirely
supported, until we have some way to override a role's method in
a consuming class.

The test changes in this patch are there to make sure downstream
code (specifically OpenTelemetry::SDK::Trace::Span) can continue
to use this approach, but should not be considered an endorsement.