Clarify in POD that the logger method is deprecated
This method was added when there was still the assumption that
we needed to provide a custom logger, rather than allow users to
provde their own.

The work to support the OTel Logs API reveals this method to be
nothing but an obstacle, since it will collide with the concept of
an OpenTelemetry::Logs::Logger, and it will confuse things when we
expose the OpenTelemetry::Logs::LoggerProvider.

All-in-all, this method provides very little benefit and obstructs
the implementation of desirable aspects of the OTel spec, so it has
to go.