Rename integration namespace to instrumentation
Writing the guides in the OpenTelemetry::Guides namespace revealed
that while the specification _always_ refers to "instrumentation
libraries", this distribution _always_ used the term "integration"
instead.

This (breaking) change renames the ::Integration namespace to more
closely align with the specification, hopefuly reducing any
confusion that might arise there. Since some code might already be
relying on that legacy namespace, this change makes efforts to
continue to support the legacy namespace, although this will likely
be removed in the future.