Remove arguments from context constructor
The OpenTelemetry::Context constructor accepted a list of values
that it would blindly set as the contents of the newly created
instance. This inadvertently allowed users to set values in the
context without using the OpenTelemetry::Context::Key objects,
and with that broke the contract of the class.

This mechanism has now been moved into a private method which should
allow most valid uses of the context class to remain unchanged.

Passing arguments to the constructor now issues a warning, which may
be removed in the future.